+++ src/features/habits/types/Frequency.ts
@@ -1,0 +1,1 @@
+export type Frequency = "Mon" | "Tue" | "Wed" | "Thu" | "Fri" | "Sat" | "Sun";
